Increased Graphics Cap!
- This commit increases the overworld graphics cap from 256 to 512.
- **YOU WILL NEED TO REBUILD TOOLS AND CLEAN BUILD THE ROM!**
- Lots of changing u8s to u16s.
- The main change is in global.fieldmap.h, where the EventObjectTemplate 
was changed to handle a u16, and the EventObject was change to handle a 
u9 for the graphicsId. (The compiler complained about making spriteId a 
u6, because it can't hold the value 64).
- All maps are affected. Mapjson will need to be recompiled, and all 
maps will need to be recompiled, or you will see Brendan everywhere.
- Old save files will load events incorrectly, but all should be fine 
with a map change.

By https://github.com/tustin2121