AI: avoid spamming Destiny Bond when caught
If the AI, while being faster than the opponent, uses Destiny Bond but it somehow survives the turn, then the optimal play is almost never to keep using it (maybe if the target cannot switch and cannot avoid choosing damaging moves). Before this commit, the AI did start spamming Destiny Bond, which the opponent can easily handle by switching around or using non-damaging moves (for example, stat up moves or Spikes). This changes make the AI to use Destiny Bond randomly if it did not work the first time. More details in the comments.