AI: tuning the switching criteria when intoxicated
If the AI chose to use Rest, Refresh, Heal Bell or Aromatherapy, then it does not need to switch out in order to prevent high Toxic damage. Otherwise, if the AI chose to use moves like Recover, then there is little point in doing so and switching out is better, even if the AI considered it a good move.