AI: make the criteria for choosing switch-ins stackable
Due to an oversight, before this commit the AI might incorrectly discard some interesting switch-ins in situations where not switching was not a horrible option (for the designed AI criteria). I am deeply sorry to all Midele Emerald players, as the AI might have been a little, if not significantly, less challenging as a consequence.