AI: control whether the opponent is switching around (#3)
This allows the AI to find out if the opponent is switching between mons that are immune or resistant to whatever the AI chooses each turn, and if so, the AI might consider to repeat the last move, in order to hit the switchin for good. The most notable effect of this addition is that it ruins the possibility of just owning the AI by using Toxic and switching around between two mons that, by covering each other's weaknesses, resist or are immune to whatever the AI uses for the other one.

The author is not liable for potential damage on a screen, a keyboard, a phone or whatever, as a consequence of this commit's effect in the sanity of the player.