AI: consider switching sooner when intoxicated
If the AI mon is intoxicated, now it might consider switching out a turn sooner than before if it is, in addition, under the effect of Leech Seed or Curse, or if it is damaged by the weather and does not have Leftovers. It might also switch out sooner if one of the two of being damaged by the weather and not having Leftovers is true, but not that likely, choosing randomly between 3 and 4 turns.

Protect in the opponent was considered, but the code was not that easy.