AI: fix bug in get_protect_count
The count is only updated when Protect/Detect/Endure is used; at that time the game checks whether the previous move was also of this type, resets the counter to 0 otherwise, and then checks whether the move is used successfully, updating the counter again (see `Cmd_setprotectlike` in `battle_script_commands.c`). Thus, the counter stays at a positive level if another move is used. As the AI uses this function to check whether it is safe to use those moves, or whether the opponent will consider using Protect for the same reason, this should return zero if the previous used move is not one of those. Otherwise, the AI refrains too much from using Protect if it was used successfully once or more and then another move was used.

While this bug is present in vanilla AI, here in Midele Emerald it affects more places; for example, this bug might make the AI explode too much against mons with Protect if the user used it successfully but not in the previous turn.