AI: better criteria for using WoW, Reflect and Light Screen
The AI used to check the types of the opposing battler in order to decide whether to use Reflect or Light Screen, when having one of them or both. Now it checks the expected category of the opponent's attacks, using the command `get_possible_categories_of_foes_attacks`, which resorts to memory and, if no memory is available, throws a reasonable guess by checking the base stats from the foe's species.

In addition, now the AI has any criteria on whether to use WoW on any opponent that does not have Wonder Guard, apart from the obvious immunity, expected Guts, and so on which make it undesirable. The AI will consider using WoW if the opponent is, or is expected to be, a physical attacker, and will mostly not use it in a special attacker, unless no good option is available.