AI: slightly discourage Thrash-like moves, fix Overheat, Superpower
Modified the criterion for encouraging moves like Overheat and Superpower: always prefer another move getting a KO if available, but it is a good option if it is the only KO-ing move. White Herb is not relevant here. Also, the same criterion can apply to Thrash and similar moves.

In addition, the code around this has been reworked and labels have been modified.