AI: Vital Throw and Revenge
The AI should not use Vital Throw or Revenge against slower mons which could OHKO the AI's. In addition, the check for a sleeping target in Revenge has been modified to be more precise, in order to not discourage Revenge against a sleeping target if the target is bound to wake up.

As a bonus, the code in Roar has been simplified.