AI: remove a thing in Sleep Talk
The AI used to consider to stay while stuck in Sleep Talk due to a choiced item if it was going to wake up immediately (and the opponent is not likely to KO the AI's mon, and the AI's mon is not the last), in order to prevent the sleep counter to reset. However, in this game, at this moment, this counter does not reset when switching out, so it makes no sense: once the mon is switched out and in, it will ìmmediately wake up and the AI knows it.