AI: Speed decrease as a secondary effect
The (programmers of the) AI forgot to take into account Mist, sub, some abilities and the possible fact that the opponent might be at -6 when using moves which decrease Speed as their 100% chance secondary effect. Now, in those circumstances, the AI considers the move as any other one with no side effect.