Add mon icon in EV changer
This reminds the player which mon is having its EVs modified. In order to accommodate for the larger icon, a new tileset has been added, with a larger background window for the mon icon than the one for items. The tilemap for the EV changer is no longer used, as the vanilla one for the shop is now used for the EV changer, but it has not been removed.