AI: check for Lock-On before raising Evasion
The check if_this_attack_might_be_the_last was omitted when evaluating Evasion-raising moves, as raising Evasion gives an opportunity to avoid attacks. However, in the rare case the mon is locked on (by Lock-On or Mind Reader), then the AI should watch out for a possible OHKO from the opponent.