Fix glitch when setting friendship for random battles
An oversight (and a silent compiler) made the game assign whatever was in position 0 in memory to the friendship of the mon. This was unnoticed, as apparently the value there is 2, pretty close to the intended 0, and therefore Frustration had almost full power, but this was catched once we inserted the code to see the power for Frustration (and Return) and running random battles in Challenge Cup mode.