AI: add macro for checking for status
This macro allows the AI to take into account that, if the target is guaranteed to wake up before the user attacks, then it might probably not matter that the target is statused. Thus, the AI might throw status to mons after two turns of Rest or after the maximum possible number of sleep turns minus 1.