AI: dampen penalty for binding moves
Now the partial trapping moves (Bind, Wrap, Whirlpool, Fire Spin...) do not receive too large a penalty for doing close to zero damage, as the secondary effect compensates a bit, unless the secondary effect will probably not happen (for being already active or if the target could have a sub or KO the user right away).

In a similar way, that penalty has also been reduced for Grav Apple and Apple Acid, as they might inflict more damage next turn or facilitate the work for other battlers, and for Midele Power.

Also, removed double super effective check which currently makes no sense, as the AI is already capable of estimating the damage and will use these moves if they are useful. Anyway, this check could make the AI easier to predict.