AI: revision of Agility
(and Rock Polish)
The AI is now less interested in using these moves when paralyzed (unless it has Shed Skin), or, when already faster, against the last mon or a mon with Encore. In addition, the criteria for using these moves when having Baton Pass are now ignored if the AI has no more available mons.