AI: consider changing sooner in some cases
Now the AI will not ignore the chance to switch out if the AI's mon will attack twice before fainting while their attacks will be 4HKO and either the target already does not have too much HP or is Leech Seeding the AI's mon.

In addition, the AI will take into account Perish Song, Yawn, or the fact that the AI is considering Explosion, when counting the available turns. This makes the AI likelier to consider switching out when yawned. Unfortunately, it is still a bit coward when considering switchins: the AI will not change unless a really good switchin is found.

Some checks were redundant and have also been removed.