AI: only encourage Perish Song if enough Wrap turns
The AI will not give more points to Perish Song if that is only because a wrapped (or similar) target which is actually going to be released soon (in less than 4 turns, even giving one turn of margin; it could do it if there were 3 turns left, but the check is not as easy to implement). This can make the AI correctly guess the number of remaining Wrap turns, but this is not as dramatic as in vanilla Emerald as, in this game, Wrap-like moves last for 5 or 6 turns (including the one where it is used).

A more reasonable alternative would be to encourage Perish Song if the target was trapped in the previous turn or in the one right before that, as that guarantees a minimum of 3 remaining turns of Wrap in this game, but the game currently provides no way to do so. Also, checking if the last move was Wrap-like and only encouraging Perish Song immediately after that is not a viable option as the mon could as well have used Wrap while already in effect, having nothing better to use.