AI: consider screens and weather when sending mons
Now, when considering sending some mon (by switching out or being forced), the AI will take into account all of these things:
 - Reflect and Light Screen are irrelevant if they are about to finish (that is, if no mon will take advantage of them), so they can then be ignored when doing calculations for evaluating how appropriate is the sent mon.
 - The same goes for the weather: if a non-permanent weather is about to finish, the AI can safely ignore it.
 - In addition, if the sent out mon has a weather ability, the AI will do the calculations with the new weather.

Note that the AI will not consider these effects to be 'about to finish' if the opponent has yet to attack during the turn (that is, if the AI is considering switching out).

These changes should make the AI a bit (if not considerably) keener on bringing mons with (non-currently-present) weather abilities (especially if an opposing weather is on), as the AI will notice that they will sometimes be sturdier and more powerful, so this can be an important step to making the AI fight weather wars (though the AI is not yet trying to keep the weather setter alive, which is the main issue preventing the AI from doing so). In addition, this should make the AI more confident about bringing mons against screen-protected opponents, more careful for the opposing situation of the screens about to be lost, and more aware of the effects of the weather being about to get to the neutral one.

There are probably more effects brought by the incoming mons' abilities that the AI still fails to take into account.