AI: use new command and macro for speed in future turns
This uses the new command in 6abe43f9500c3830f8c4d38aabf7eaa1c710e730. Most notably, the AI will not use Agility for nothing if the target will still be faster and 2HKO the AI's mon, which is one of the most notable weaknesses of the AI before this commit.