AI: do not forget the suitable mon found when considering switching out (#15)
Commit 556f00d7940 added a way for the AI to switch out when all the moves had a low score, which comes itself from DizzyEggg/pokeemerald@272fa07799. It has been a really long way from this, and now the AI takes advantage of this for many other criteria for switching out after evaluating the moves, choosing suitable mons with different criteria depending on the situation. This has been possible thanks to that.

However, there is a small problem. The way commit DizzyEggg/pokeemerald@272fa07799 is designed, the most suitable mon found is completely ignored. The most suitable mon has to be chosen again in `AI_TrySwitchOrUseItem`. Note that it is not stored after calling `GetMostSuitableMonToSwitchInto` in `battle_ai_script_commands.c`; only the fact that there were more than zero suitable mons is stored in `switchMon`.

In addition to a waste of CPU, this can affect our current criteria for switching out, as they depend on the direness of the situation, and `AI_TrySwitchOrUseItem` always considers that not changing is _unacceptable_. It mostly does not, as the criteria are similar except some discard mons in not too bad situations while in other cases the criteria cannot discard all mons (e.g. when not changing is impossible), but maybe it will in the future.

In addition, if the AI somehow controls the player too (yes, I have been testing that 👀), the value in `switchMon` can be overwritten before the time the AI reconsiders if it has to switch out (it is for example unset when starting to give scores to the moves). This can force the AI to recalculate _everything_, including evaluating the attacks. This does not lead to a loop, as at least one of the AIs gets it right, but it is even more wasteful. In addition, this could be a problem should we allow the AI of the player's partner to consider switching out (however, double battles are currently pretty much ignored throughout the AI).

The solution is to actually use a field in `gBattleStruct` which fits this purpose. It is preserved between AI evaluation steps (unlike `AI_THINKING_STRUCT`) but is unset before the beginning of the turn, and it has an entry for each of the up to 4 mons in the field.

In addition, a small rearrangement of the criteria for switching out has been done, and the function of checking if a switching was decided has been renamed in order to not suggest that there are no more criteria than the one of switching out because of bad scores in the moves.