AI: take paralysis into account when checking if target will be faster after the current effect
If the AI's mon is about to heal itself from paralysis, then it should do the speed comparison ignoring the paralysis (assuming the paralysis does not prevent the AI from attacking). If the target is about to be paralyzed, then the AI can consider that paralysis for the next turn (assuming it hits).