AI: tune down stat ups when not causing enough damage
Now, when considering the moves that increase Attack and/or Special Attack (this includes DD, QD, CM...) without having Baton Pass, the AI will not be too eager to use the move due to having a sash or expecting to be safe to use the move, if the AI will not cause too much damage to the opponent. This prevents the AI from uselessly boosting against a powerful opponent that will take negligible damage even after the stat up, and additionally the AI will be able to consider other (non-damaging) good moves, if any, before starting to spam stat-ups if it will take long in order to notably damage the target.

This might make the AI use recovery moves too soon for sets like the one for Reuniclus you are thinking of right now. This has to be tested.