AI: take into account Perish Song in more scenarios
Many moves make close to zero sense if the user is about to get fainted through Perish Song. For some other moves, the same is true for the target. Not immediate Perish Song has also been added to be considered for more moves.

In addition, the penalties for Ingrain have been increased.