AI: criteria for switching out while having Truant
Now the AI considers whether to withdraw its mon with Truant for being particularly vulnerable against the opponent (Protect...) before evaluating how convenient the mon's moves are. In addition, the AI might switch out the mon with Truant if it is about to loaf around and no move has at least 103 points, as that means that the Truant mon is unlikely to successfully use a OHKOing move, but only if a good switchin is found.