AI: Destiny Bond against locked mons
If the target is slower and in a multiturn attack, then Destiny Bond
should not be used if the mon will clearly survive next turn; otherwise,
it depends on the available mons in each team.

In addition, various penalties for Destiny Bond have been increased so
the AI will prefer switching out or using attacks that do at least some
damage. The final bonuses for low HP have also been reduced.