AI: prevent the AI from considering eggs
Under normal circumstances, the AI does not control teams where more
than zero mons are eggs. However, it was easy to fix the AI in case this
happens to be false, somehow :eyes:

I did not check whether the game gets stuck when trying to send an Egg
and whether is feasible to not discard the Egg one way or another, but
the Pomeg glitch suggests that the AI might in fact "successfully" send
an Egg.