AI: check weather at the end of the turn
The current weather does not matter from the point of view of taking (or
healing) damage at the end of the turn if it is expected to finish at
the end of the current turn, so a command has been made for checking
just that. This is also taken into account in another place that is not
in the AI scripts.