AI: Endure
The AI will now avoid Endure if Sturdy or sash would do the same or if
the opponent is known to have no damaging moves. In addition, the AI
will better use Endure depending on the expected damage from the target,
and will not discard Endure when having really low HP if the target is
being afflicted by damaging statuses and cannot heal.