AI: Steel mons with Magnet Pull can switch out
The vanilla AI had a glitch where it checked whether any mon on the
field has Magnet Pull, including the current mon, in order to decide
whether it can be switched out (and yes, this means that the vanilla AI
will let their Magnetons be afflicted by Perish Song). When checking for
switchins in the other part of the code (after evaluating the moves),
the same glitch was carried.