AI: Pursuit
The AI will not be particularly interested in Pursuit if the target
cannot switch out. Otherwise, if the target just used a move that has no
effect on the current mon, this means that the current mon just entered
the field, so it has a chance to make the opponent flee, especially if
it turns out to be choiced. (Only some immunities that are common among
Pursuit users are checked)

Too bad Pursuit is mostly ignored in Midele Emerald.