AI: take into account imminent Perish Song for more moves
The ones defense-related are not taken into account since there is a
nonzero chance that they might actually be useful (so that the opponent
faints at the same time as the user, or, if it has some status, before
the user), so this covers (phys/sp)Attack, Speed, Accuracy-boosting
moves, Block, Foresight and Lock-On. While Perish Song is also taken
into account in AI_CV for most of them, if not all, a greater penalty is
reasonable (there it is usually -5 since it mostly considers whether
Perish Song is active, not necessarily imminent).