AI: criteria for switching instead of repeating sub
The AI will now attempt to switch out instead of repeating Substitute
when intoxicated, cursed or under Perish Song, even if some conditions
that made reasonable spamming sub (pinch berries, an opponent losing HP)
are fulfilled. In addition, the AI will now try to switch out when
repeatedly using SubProtect (in the turn of sub), using the same
criteria as when repeating sub.