AI: Encore
It turns out that Encore cannot Encore Encore, so Encore should be
removed from the list of moves encouraged to Encore. Now Encore is
avoided if the target used Encore, Struggle, Mirror Move, Sketch or a
move that the target will be unable to keep using due to not having PP.
On the other hand, the AI will encourage Encore if the target will
probably have to use Struggle due to the Encore (that is, if other moves
like Taunt, Torment or possibly Disable prevent the move from being
repeated).