AI: semi-invulnerability moves
- The AI will now avoid exploding against targets that will be, or might
be, in a semi-invulnerable status.
- The AI will now consider semi-invulnerable moves as regular two-turn
attacks if No Guard or Lock-On apply.
- The AI will now try to use semi-invulnerable moves against incoming
semi-invulnerable moves if the target is slower and the received damage
is considerable, and avoid doing so if the target is faster, since it
could repeat it and no damage would be dealt.
- The AI will also be less likely to use Protect against targets that
use a semi-invulnerable move that causes pathetic damage.

In addition, the penalty for using Explosion against the last mon of the
opponent when it is under Perish Song has been increased.