AI: discourage moves against opponents with Sub
The game had a check for some moves that currently makes little sense
for most of them, since the AI now correctly estimates the damage of
moves such as Return, Seismic Toss, Flail, or Low Kick, and the AI makes
a reasonable assessment on whether to use Focus Punch. For the others
(as well as Overheat and similar ones), now the AI checks whether the
target could spam Substitute in order to defend itself from the attack.
This prevents the AI (unless the other moves are terrible) from using
moves such as Overheat, Superpower, Outrage, Razor Wind or Counter
against targets that might use Sub enough times to defeat the move; in
particular, the AI will no longer be eager to get an OHKO with such
moves.