AI: Fake Out
Now the AI will not be that eager using Fake Out when choiced and being
the last mon if the KO is possible but not guaranteed. In addition, now
Fake Out does not get penalty for not being the most damaging move,
unless the target has a sub or might have Inner Focus or Shield Dust and
the target is faster and would KO before other attacks (and in that
case, it now gets the same bonus as usual).