AI: do not switch for bad moves in more cases
Now the AI will not switch out as a consequence of only having poor
moves when at +6 Evasion if additionally the target is under Leech Seed.
Also, the AI will not allow (almost) any mon to be switched in as a
consequence of having terrible moves if any Defense or the Evasion is
over +3; only reasonable switchins will be allowed in that case.