AI: AI_CV_Wish
Now the AI should mostly not spam Wish when nothing good is available;
especially if the opponent has no powerful moves against the user, or if
the opponent does but the AI will not have time to benefit from Wish by
at least switching out (e.g. expects a 2HKO when trapped or being the
last mon).