AI: Endeavor
- The AI now knows the damage of Endeavor if it will be faster than the
opponent or has a substitute. This itself greatly improves the AI's
capability to decide whether to use Endeavor when faster. The points for
this move when faster have been also tweaked.
- The AI takes into account incoming Perish Song, considering its use at
the last turn if it cannot switch out in order to hit the switch-in.
- Major changes in the logic of Endeavor when slower, trying to
anticipate whether the AI will lose enough HP in order to use Endeavor.

Also, removed some unused and repeated code.